requirements.txt
streamlit
groq
requests
httpx
python-dotenv
streamlit-cookies-manager==0.2.0
pandas
fastapi
uvicorn
resend